+++ baobab/lims/tests/variables.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of Bika LIMS
+#
+# Copyright 2011-2017 by it's authors.
+# Some rights reserved. See LICENSE.txt, AUTHORS.txt.
+
+"""Variables for lims/tests/robots/*.robot"""
+
+PLONEURL = "http://localhost:55001/plone"
+SELENIUM_SPEED = 0.1
